requirements.txt
#pip install -r requirements.txt


fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose
python-jose[cryptography]

